feat(employe): add employee context
